requirements.txt
ipython==8.12.3
matplotlib==3.9.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
pycocotools==2.0.8
pytorch-lightning==2.3.3
PyYAML==6.0.2
scikit_learn==1.5.2
Shapely==2.0.6
skimage==0.0
timm==1.0.9
torch==2.4.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.65.0
ultralytics==8.2.98

json_tricks==3.17.3
mmcv==2.2.0
mmdet==3.3.0
mmengine==0.10.5
mmpose==1.3.2

# Core packages
numpy==1.24.4
pandas==2.0.3
matplotlib==3.9.2
scikit-learn==1.5.2
torch==2.4.0
torchvision==0.19.1
pytorch-lightning==2.4.0
torchmetrics==1.4.2
timm==1.0.9
tqdm==4.66.5
pillow==10.4.0
opencv-python==4.10.0.84
ffmpeg==6.1.1

# Image processing
pycocotools==2.0.8
imageio==2.34.0
scikit-image==0.24.0
seaborn==0.13.2
shapely==2.0.5
xtcocotools==1.14.3
albumentations==1.2.0

# Utilities
PyYAML==6.0.2
requests==2.28.2
filelock==3.14.0
fsspec==2024.6.1
six==1.16.0
typing-extensions==4.12.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pyzmq==25.1.2
psutil==5.9.8
ipykernel==6.29.3
ipywidgets==8.1.2
traitlets==5.14.2
pickleshare==0.7.5
setuptools==60.2.0
pip==24.2

# Deep learning and transformers
transformers==4.44.0
datasets==2.14.4
huggingface-hub==0.24.6
grad-cam==1.5.2
ultralytics==8.2.98
ttach==0.0.3
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.4

# OpenMMLab specific utilities
addict==2.4.0
chumpy==0.70
json-tricks==3.17.3
markdown==3.7
markdown-it-py==3.0.0
openmim==0.3.9
tabulate==0.9.0
terminaltables==3.1.10
yapf==0.40.2